Obsolete cipher_state_preprocessor removed #18
